// src/generated/resolver-types.ts
export type Maybe<T> = T | null;
// ...
export type RequireFields<T, K extends keyof T> = Omit<T, K> & { [P in K]-?: NonNullable<T[P]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
};

export type Hero = {
  __typename?: 'Hero';
  genres: Array<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  language: Scalars['String']['output'];
  matchPercentage: Scalars['String']['output'];
  overview: Scalars['String']['output'];
  poster?: Maybe<Scalars['String']['output']>;
  releaseDate?: Maybe<Scalars['String']['output']>;
  title: Scalars['String']['output'];
  trailer?: Maybe<Scalars['String']['output']>;
};


export type HeroGenresArgs = {
  media: Media;
};


export type HeroTrailerArgs = {
  media: Media;
};

export enum Media {
  Movie = 'movie',
  Tv = 'tv'
}

export type Query = {
  __typename?: 'Query';
  getHero: Hero;
};


export type QueryGetHeroArgs = {
  media: Media;
};



export type ResolverTypeWrapper<T> = Promise<T> | T;


export type ResolverWithResolve<TResult, TParent, TContext, TArgs> = {
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> = ResolverFn<TResult, TParent, TContext, TArgs> | ResolverWithResolve<TResult, TParent, TContext, TArgs>;
// ...
export type IsTypeOfResolverFn<T = {}, TContext = {}> = (obj: T, context: TContext, info: GraphQLResolveInfo) => boolean | Promise<boolean>;

export type NextResolverFn<T> = () => Promise<T>;

export type DirectiveResolverFn<TResult = {}, TParent = {}, TContext = {}, TArgs = {}> = (
  next: NextResolverFn<TResult>,
  parent: TParent,
  args: TArgs,
  context: TContext,
  info: GraphQLResolveInfo
) => TResult | Promise<TResult>;



/** Mapping between all available schema types and the resolvers types */
export type ResolversTypes = {
  Boolean: ResolverTypeWrapper<Scalars['Boolean']['output']>;
  Hero: ResolverTypeWrapper<Show>;
  ID: ResolverTypeWrapper<Scalars['ID']['output']>;
  Media: Media;
  Query: ResolverTypeWrapper<{}>;
  String: ResolverTypeWrapper<Scalars['String']['output']>;
};

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = {
  Boolean: Scalars['Boolean']['output'];
  Hero: Show;
  ID: Scalars['ID']['output'];
  Query: {};
  String: Scalars['String']['output'];
};

export type HeroResolvers<ContextType = any, ParentType extends ResolversParentTypes['Hero'] = ResolversParentTypes['Hero']> = {
  genres?: Resolver<Array<ResolversTypes['String']>, ParentType, ContextType, RequireFields<HeroGenresArgs, 'media'>>;
  id?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
  language?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  matchPercentage?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  overview?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  poster?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  releaseDate?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  title?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  trailer?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType, RequireFields<HeroTrailerArgs, 'media'>>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type QueryResolvers<ContextType = any, ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query']> = {
  getHero?: Resolver<ResolversTypes['Hero'], ParentType, ContextType, RequireFields<QueryGetHeroArgs, 'media'>>;
};

export type Resolvers<ContextType = any> = {
  Hero?: HeroResolvers<ContextType>;
  Query?: QueryResolvers<ContextType>;
};